+++ src/debug.c
@@ -166,6 +166,7 @@ int desensamblar_instruccion(const Chunk *c, int offset, FILE *out) {
         case OP_LISTA_AGREGAR:   return instruccion_simple("OP_LISTA_AGREGAR", offset, out);
         case OP_LISTA_EXTENDER:  return instruccion_simple("OP_LISTA_EXTENDER", offset, out);
         case OP_LLAMAR_SPREAD:   return instruccion_simple("OP_LLAMAR_SPREAD", offset, out);
+        case OP_CONJUNTO_AGREGAR: return instruccion_simple("OP_CONJUNTO_AGREGAR", offset, out);
         case OP_LLAMAR_KW: {
             /* 2 bytes: n_pos, n_kw. */
             uint8_t np = c->codigo[offset + 1];